Move CoreOps fully into package
