--- modules/coreops/__init__.py
@@ -1,1 +1,0 @@
-# CoreOps module namespace (Phase 1 commands live here).
