--- modules/coreops/cog.py
@@ -1,10 +1,0 @@
-"""CoreOps module loader."""
-
-from __future__ import annotations
-
-
-async def setup(bot):
-    from c1c_coreops.cog import CoreOpsCog
-
-    await bot.add_cog(CoreOpsCog(bot))
-
