--- modules/coreops/commands/__init__.py
@@ -1,3 +1,0 @@
-"""CoreOps command package."""
-
-__all__ = ["reload"]
