+++ packages/c1c-coreops/src/c1c_coreops/cron_summary.py
@@ -1,3 +1,5 @@
+"""Daily summary helpers for CoreOps cron telemetry."""
+
 from __future__ import annotations
 
 import datetime as dt
